23.06.2026 Perenos

+++ code/modules/spells/pantheon/inhumen/zizo.dm
@@ -188,6 +188,8 @@
 		qdel(src)
 		return BULLET_ACT_BLOCK
 
+	if(out_of_effective_range())
+		return
 	try_embed_target(L)
 	qdel(src)
